Sửa theo bản review

+++ src/main/java/com/example/trackingorders/dto/response/ReturnSummaryResponse.java
@@ -6,13 +6,14 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 @Getter
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
 public class ReturnSummaryResponse implements Serializable {
-    private Integer activeReturns;
-    private Integer awaitingInspection ;
-    private Integer totalRefunds ;
+    private Long activeReturns;
+    private Long awaitingInspection ;
+    private BigDecimal totalRefunds ;
 }
